Fix static analysis issues

+++ src/eglfs_webos/deviceintegration/eglfs_starfish/eglfsstarfishintegration.h
@@ -82,7 +82,7 @@ public:
                                                       drmModeConnectorPtr connector,
                                                       ScreenInfo *vinfo,
                                                       const QString& connectorName,
-                                                      int crtc,
+                                                      size_t crtc,
                                                       int selected_mode,
                                                       QList<drmModeModeInfo> modes);
